kripto_app/requirements.txt
streamlit
psycopg2-binary
pycryptodome
pillow
